[FIX] 오타 수정, 배포 시 나는 오류 수정 #34

See merge request s12-webmobile2-sub1/S12P11A408!39

+++ saju-backend/src/main/java/com/saju/sajubackend/api/couple/dto/CoupleResponseDto.java
@@ -28,7 +28,7 @@ public record CoupleResponseDto(
             Long memberId,
             String nickname,
             String profileImage,
-            Integer region,
+            Long region,
             int age,
             String celestialStem
     ) {
