+++ saju-backend/src/main/java/com/saju/sajubackend/api/filter/dto/MemberProfileResponse.java
@@ -10,7 +10,7 @@ public class MemberProfileResponse {
     private String nickname;
     private String gender;
     private String profileImage;
-    private int cityCode;
+    private long cityCode;
     private String religion;
     private int age;
     private int height;
